Deconfuse some heavy use of the preprocessor.
One #if...#endif started in one function's body and ended in
another's, which is worth remarking on in comments at both ends.

A later #if...#else... opened blocks in both halves, that were closed
just after their #endif, leading to simplistic brace-matching failure;
so move that closing brace to inside both halves, so each half is
brace-balanced and the function, as a whole, brace-balances simply.

Commented on each #endif that was distant from its #if.

Change-Id: I74d97b6dc4b291ec05053d299077a61cecf41ca1
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>