Support arbitrary strides in the QPlatformBackingStore compositor
Otherwise platform backingstores providing QImages with extra pixels
would break horribly. The only such platform is wayland for now.

Previously the need to support this case was masked by the RGBA8888
image conversion that happened always with wayland due to its ARGB32_Pre
format. With the recent improvements this is not done anymore and so
the problems became apparent.

Task-number: QTBUG-50894
Change-Id: I27d7a1c8e25d152ca1227af1e2c38f7d4b6acbab
Reviewed-by: Paul Olav Tvete <paul.tvete@theqtcompany.com>